Fix tmux missing plugins.Work on rubocop

+++ tmux/install.sh
@@ -2,4 +2,8 @@
 
 set -e
 
-ln -sf "${DOTFILES_LOCATION}/tmux/.tmux.conf" "${HOME}/.tmux.conf"
+if ! [ -e $HOME/.tmux.conf ]; then
+    ln -sf "${DOTFILES_LOCATION}/tmux/.tmux.conf" "${HOME}/.tmux.conf"
+    git clone --depth 1 https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+    echo "▶️  Tmux setup"
+fi
